requirements.txt
pymysql
qiniu
celery
redis
django-cors-headers

eventlet